requirements.txt
discord.py==2.6.4
PyYAML==6.0.3
